auto-inventory-ai-frontend-main/python_agent_service/requirements.txt
fastapi
uvicorn
google-genai
pymongo
python-dotenv
pydantic
motor
